Payment + PDF receipt
Generate payments and download PDF receipts

+++ MedicalManagementSystem/src/main/java/com/srh/medicalmanagementsystem/dao/PaymentRepository.java
@@ -3,6 +3,5 @@ package com.srh.medicalmanagementsystem.dao;
 import com.srh.medicalmanagementsystem.entity.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentRepository extends JpaRepository<Payment, Integer> {
+public interface PaymentRepository extends JpaRepository<Payment, Long> {
 }
-
